+++ src/data-browser/constants/statePop.js
@@ -7,6 +7,7 @@ import populations2021 from './populationByGeography/statePop2021.json'
 import populations2022 from './populationByGeography/statePop2022.json'
 import populations2023 from './populationByGeography/statePop2023.json'
 import populations2024 from './populationByGeography/statePop2024.json'
+import populations2025 from './populationByGeography/statePop2025.json'
 
 export default {
   2017: populations2017,
@@ -17,4 +18,5 @@ export default {
   2022: populations2022,
   2023: populations2023,
   2024: populations2024,
+  2025: populations2025,
 }