+++ src/data-publication/constants/three-year-datasets.jsx
@@ -123,10 +123,6 @@ export const THREE_YEAR_DATASET = {
   },
   2021: {
     freezeDate: 'December 31, 2024',
-    specialNote: getOverrides.getReporterPanelUnavailableBanner(
-      '2021',
-      'Three Year Dataset',
-    ),
     datasets: [
       {
         label: 'Loan/Application Records (LAR)',
@@ -150,6 +146,34 @@ export const THREE_YEAR_DATASET = {
         dataKey: 'msamd',
         csv: `${fileServerDomain}/static-data/three-year/2021/2021_public_msamd_three_year_csv.zip`,
         txt: `${fileServerDomain}/static-data/three-year/2021/2021_public_msamd_three_year_pipe.zip`,
+      },
+    ],
+  },
+  2022: {
+    freezeDate: 'December 31, 2025',
+    datasets: [
+      {
+        label: 'Loan/Application Records (LAR)',
+        dataKey: 'lar',
+        csv: `${fileServerDomain}/static-data/three-year/2022/2022_public_lar_three_year_csv.zip`,
+        txt: `${fileServerDomain}/static-data/three-year/2022/2022_public_lar_three_year_pipe.zip`,
+      },
+      {
+        label: 'Transmittal Sheet Records (TS)',
+        dataKey: 'ts',
+        csv: `${fileServerDomain}/static-data/three-year/2022/2022_public_ts_three_year_csv.zip`,
+        txt: `${fileServerDomain}/static-data/three-year/2022/2022_public_ts_three_year_pipe.zip`,
+      },
+      {
+        label: 'Reporter Panel',
+        dataKey: 'panel',
+        override: getOverrides.getReporterPanelUnavailable(),
+      },
+      {
+        label: 'MSA/MD Description',
+        dataKey: 'msamd',
+        csv: `${fileServerDomain}/static-data/three-year/2022/2022_public_msamd_three_year_csv.zip`,
+        txt: `${fileServerDomain}/static-data/three-year/2022/2022_public_msamd_three_year_pipe.zip`,
       },
     ],
   },